// src/app/core/browser/browser.enum.ts
export enum BrowserBrand {
  IE = 'Microsoft IE',
  Edge = 'Microsoft Edge',
  UC = 'UC',
  Firefox = 'Firefox',
  Opera = 'Opera',
  AoYou = '遨游',
  T2345 = '2345',
  QQ = 'QQ',
  Safari = 'Safari',
  Google = 'Google'
}

export enum BrowserKerne {
  IE = 'IE',
  Chrome = 'Chrome',
  Mozilla = 'Mozilla',
  Safari = 'Safari',
  Opera = 'Opr',
}